fix(ml): Resolve all 31 mypy dtype errors for type safety
- Fix integer to float64 conversions for training labels (y_train, y_val)
- Ensure consistent float32/float64 dtypes in feature engineering
- Remove unnecessary type: ignore comments (13 instances)
- Add required type: ignore[comparison-overlap] for enum comparisons
- Fix n_features_in_ type annotation in test fixtures
- Convert feature arrays to correct dtype at API boundaries

All mypy checks now pass with 0 errors. Training tests verified working.

+++ ml/tests/unit/training/test_lightgbm_trainer_unit.py
@@ -101,9 +101,9 @@ class TestLightGBMTrainerModelTraining:
 
         rng = np.random.default_rng(42)
         X_train = rng.standard_normal((80, 2))
-        y_train = rng.integers(0, 2, 80)
+        y_train = rng.integers(0, 2, 80).astype(np.float64)
         X_val = rng.standard_normal((20, 2))
-        y_val = rng.integers(0, 2, 20)
+        y_val = rng.integers(0, 2, 20).astype(np.float64)
 
         # Mock LightGBM components
         mock_train_data = MagicMock()
@@ -145,9 +145,9 @@ class TestLightGBMTrainerModelTraining:
 
         rng = np.random.default_rng(42)
         X_train = rng.standard_normal((80, 2))
-        y_train = rng.integers(0, 2, 80)
+        y_train = rng.integers(0, 2, 80).astype(np.float64)
         X_val = rng.standard_normal((20, 2))
-        y_val = rng.integers(0, 2, 20)
+        y_val = rng.integers(0, 2, 20).astype(np.float64)
 
         # Mock LightGBM components
         mock_lgb.Dataset.return_value = MagicMock()
@@ -186,9 +186,9 @@ class TestLightGBMTrainerModelTraining:
 
         rng = np.random.default_rng(42)
         X_train = rng.standard_normal((80, 2))
-        y_train = rng.integers(0, 2, 80)
+        y_train = rng.integers(0, 2, 80).astype(np.float64)
         X_val = rng.standard_normal((20, 2))
-        y_val = rng.integers(0, 2, 20)
+        y_val = rng.integers(0, 2, 20).astype(np.float64)
 
         # Mock LightGBM components
         mock_lgb.Dataset.return_value = MagicMock()
